Update navier_stokes_cleaned.py
Added the function supernova sputtering. For a given x and specie u, it calculates the added/destroyed mass for the nearest dust particles and the gas particles that intersect that dust particle. It calculates these particles' new masses and the new specie fractions of u that they contain, and updates them.